requirements.txt
numpy
trimesh
Pillow
pytest
scipy
